+++ CHANGES.rst
@@ -4,6 +4,7 @@
 * ``sqlalchemy.IdMixin`` now supports UUID primary keys.
 * Deprecated ``add_and_commit`` in favour of ``failsafe_add``.
 * New utils: ``uuid2buid`` and ``buid2uuid``.
+* Removed ``timestamp_columns`` (was deprecated in 0.4.3).
 
 0.5.0
 -----